backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
Pillow==10.4.0
SQLAlchemy==2.0.35
requests==2.32.3
python-multipart==0.0.9
pydantic==2.9.2
aiofiles==24.1.0
